requirements.txt
bs4
pillow
quart
python-resize-image
Wand
